Confirmation dialogue on resource delete

+++ metadata-app/src/components/ResourceUpload/index.js
@@ -1,11 +1,14 @@
 import React, { useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
 
+import { Dialog } from '@cmsgov/design-system';
 import Link from '../Link';
 import Radio from '../Radio';
 import WrappedField from '../WrappedField';
 import resourceFormats from './resource_formats.json';
 import { RESOURCE_URL_TYPES } from '../../api';
+
+import '@cmsgov/design-system/dist/css/index.css';
 
 const sortedResourceFormats = resourceFormats.sort((a, b) => {
   const labelA = a.label.toUpperCase();
@@ -36,6 +39,7 @@ const ResourceUpload = (props) => {
   const { url, name, description, mimetype, format } = resource;
 
   const [uploadDataFileIsActive, setUploadDataFileActive] = useState(false);
+  const [shouldShowModal, setShowModal] = useState(false);
 
   // Detect when url type changes
   useEffect(() => {
@@ -64,6 +68,14 @@ const ResourceUpload = (props) => {
     setFieldValue('resource.upload', event.currentTarget.files[0]);
   };
 
+  const hideModal = () => {
+    setShowModal(false);
+  };
+
+  const showModal = (resourceName) => {
+    setShowModal(resourceName);
+  };
+
   const editResource = (res) => {
     setFieldValue('resourceAction', 'edit');
     setUploadDataFileActive(false);
@@ -83,7 +95,7 @@ const ResourceUpload = (props) => {
   };
 
   return (
-    <div className="usa-form-custom">
+    <div className="usa-form-custom" id="resource-upload">
       <section id="section-basic-mega-menu" className="site-component-section">
         <h1 className="usite-page-title" id="basic-mega-menu">
           Resource Upload
@@ -114,14 +126,43 @@ const ResourceUpload = (props) => {
                   <div className="resource-item-name">{res.name}</div>
                   <button
                     id={`delete-${res.name}`}
-                    onClick={() => {
-                      deleteResource(res);
-                    }}
+                    onClick={() => showModal(res.name)}
                     type="button"
                     className="usa-button--unstyled resource-action-button"
                   >
                     Delete
                   </button>
+                  {shouldShowModal === res.name && (
+                    <Dialog
+                      onExit={() => hideModal()}
+                      getApplicationNode={() => document.getElementById('resource-upload')}
+                      heading="Please Confirm Action"
+                      underlayClickExits
+                      actions={[
+                        <button
+                          type="button"
+                          className="ds-c-button ds-c-button--primary"
+                          key="primary"
+                          onClick={() => {
+                            deleteResource(res);
+                            hideModal();
+                          }}
+                        >
+                          Delete
+                        </button>,
+                        <button
+                          type="button"
+                          className="ds-c-button ds-c-button--transparent"
+                          key="cancel"
+                          onClick={() => hideModal()}
+                        >
+                          Cancel
+                        </button>,
+                      ]}
+                    >
+                      Are you sure you want to delete this resource?
+                    </Dialog>
+                  )}
                   <button
                     id={`edit-${res.name}`}
                     onClick={() => {
